% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.files.r
\name{cut.files}
\alias{cut.files}
\title{cut.files}
\usage{
\method{cut}{files}(files, d.type = "GPS", radius)
}
\arguments{
\item{files}{a vector of the file index.}

\item{d.type}{}
}
\description{
cut.files
}
