% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.chooser.r
\name{files.chooser}
\alias{files.chooser}
\title{Choose only the files you want}
\usage{
files.chooser(output = "filenames", folder = "GPS", year = "all",
  group.num = "all", g_s = "all", f_u = "all", site = "all",
  miles = "all", direction = "all", flight.nums = "all",
  pigeon.nums = "all", release.of.day = "all")
}
\arguments{
\item{output}{"numbers" or "filenames". Filenames probably more useful. We'll see.}

\item{folder}{"GPS" or "ACC"}

\item{year}{Year 1 or Year 2 of Phd}

\item{group.num}{c(1,2) etc. or "all"}

\item{g_s}{group, solo or all = "g", "s" or "all"}

\item{f_u}{familiar, unfamiliar or all = "f", "u" or "all"}

\item{site}{Site 1,2,3,4,5,6,7,8,9,10 or "all"}

\item{miles}{4, 8 or "all"}

\item{direction}{N, NE, E, SE, S, SW, W, NW or "all"}

\item{flight.nums}{1:12 or "all"}

\item{pigeon.nums}{or c(32,39,40) etc.}

\item{release.of.day}{1,2, 3 , c(2,3) or "all"}
}
\description{
other packages used in this function: stringr
}
