% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dist.R
\name{get_dist}
\alias{get_dist}
\title{Get the distance between two GPS vectors}
\usage{
get_dist(lon1, lat1, lon2, lat2, hz, method = c("speed", "distance"))
}
\arguments{
\item{lon1}{longtude vector for individual 1 or if method is "speed" then t = 1}

\item{lat1}{latitude vector for individual 1 or if method is "speed" then t = 1}

\item{lon2}{longtude vector for individual 2}

\item{lat2}{longtude vector for individual 2}

\item{hz}{Only when method is "speed". This is how many locations per second}

\item{method}{When method = speed, no need for lat2 and lon2}
}
\value{
The distance between the input locations in metres
}
\description{
Takes two latitudes / longitudes and calculates distance in metres
}
