% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_head.R
\name{get_head}
\alias{get_head}
\title{get_head}
\usage{
get_head(lon1, lat1, lon2, lat2, indivs = 1)
}
\arguments{
\item{lon1}{lonitude of individual 1}

\item{lat1}{latitude of individual 1}

\item{lon2}{lonitude of individual 2}

\item{lat2}{latitude of individual 2}

\item{indivs}{If indivs = 1 you will not need lat2 and lon2}
}
\description{
get the heading between two GPS coordinates
}
