% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricF.R
\name{metricF}
\alias{metricF}
\title{metricF}
\usage{
metricF(accuracy = T, speed = T, time.taken = T, min.speed = 5,
  hz = 5, files = files.chooser(folder = "mGPS"))
}
\arguments{
\item{accuracy}{the accuracy of the path compared to the beeline}

\item{time.taken}{The number of rows of data divided by the hz of the logger}

\item{files}{Choose the files you want to perform functions on. USe files.chooser( g_s = "g") to get groups for example. Do not change folder = "mGPS"}
}
\description{
A function to produce metrics for each flight. eg. the accuracy of the flight path. Works in the mGPS folder.
}
