% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_row.R
\name{metrics_row}
\alias{metrics_row}
\title{metrics_row}
\usage{
metrics_row(file, input = c("GPS", "ACC", "ufi&id"), output = c("row_num",
  "pigeon", "group.num", "f_u", "site", "g_s", "dir", "miles", "flight",
  "accuracy", "mean.speed", "var.speed", "total.dist", "time.taken",
  "time.stationary"))
}
\description{
Which row in the metricws dataframe corresponds to our data?
}
