% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mGPS.R
\name{mGPS}
\alias{mGPS}
\title{mGPS}
\usage{
mGPS(lat_lon2numeric = T, correct_time = T, correct_lon = T,
  get_speed = T, get_heading = T, cut_start = T, cut_end = T,
  cut_radius = 200, files = files.chooser(output = "filenames"), save = T,
  return_data = F)
}
\arguments{
\item{lat_lon2numeric}{necessary when NAs present in data}

\item{correct_time}{Sometimes the GPS saves random additional files. This will get rid of those.}

\item{correct_lon}{Multiply the output data by negative 1. Longitude data from my GPS is on the wrong side of zero.}

\item{get_speed}{Speed in m/s}

\item{get_heading}{Heading for each individual's trace.}

\item{cut_radius}{is the cut radius in metres}

\item{cut_files}{Cut the files at a radius from the release site and home}
}
\description{
A function which modifies raw GPS files and outputs in the mGPS folder
}
