% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{plot_map}
\usage{
plot_map(data, type = c("solo", "group"), image.or.vid = "video",
  tail.size = 20, frame.size = 1, wd = getwd(), plot_name = "my_plot",
  lines.or.points = "lines", col = "red", lwd = 2, file.to.folder = T,
  maptype = "satellite", zoom = 19)
}
\arguments{
\item{data}{data array with dimensions time (rows), longitude / latitude (in that order; columns) & individual (3rd dimension). No third dimension necessary for solo plots.}

\item{type}{Solo or Group plots.}

\item{image.or.vid}{"image" or "video" (sequence of images) to be produced}

\item{tail.size}{animals represented by current position and the previous n positions (video only)}

\item{frame.size}{plot every row of data, or every 1/frame size for shorter running speeds and output files (video only)}

\item{wd}{Specify the working directory that you would like to save the plots into.}

\item{lines.or.points}{On the plotted map, animal trajectory represented by either "lines" or "points"}

\item{col}{Colour of plotted trajectory (solo only)}

\item{lwd}{Thickness of line}

\item{file.to.folder}{TRUE if you want to make videos when happy with other params, but for testing videos set to FALSE}

\item{maptype}{"satellite" is default, though check ?get_map in the ggmap package for more options}

\item{zoom}{zoom = 19 is good for video. Other plot types are worth playing around with, try 13 for approx 5 km square}
}
\description{
A handy and interactive plotter for solo or group GPS tracks. Uses ggmaps to produce images or videos of animal trajectories on a googlemaps background.
}
