% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d.R
\name{plot_pidge_3d}
\alias{plot_pidge_3d}
\title{plot_pidge_3d}
\usage{
plot_pidge_3d(d3, box.size = 40, tail.size = 40, frame.size = 2,
  prev.wd = getwd(), lwd = 2, wd, grid = "equal", axes = F,
  axis.lab = F)
}
\arguments{
\item{d3}{an array of flight data structured in three dimensions 1.time, 2.(x,y,z), 3.id}

\item{box.size}{size of the box around the data}

\item{tail.size}{the size of the pigeons tails}

\item{frame.size}{to subset the dataset (larger numbers use less memory)}
}
\description{
plot_pidge_3d
}
