% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traj.R
\name{plot_traj}
\alias{plot_traj}
\title{plot_traj}
\usage{
plot_traj(data, type = c("solo", "group"), snap.or.vid = "video",
  borders = 800, tail.size = 20, frame.size = 1, grid.width = 10,
  png = T, hz = 5, lines.or.points = "lines", f_u = "all",
  plot.links = T, d2c.plot = T, window = c("full", "tail.size"))
}
\description{
A handy and interactive plotter for solo or group GPS tracks. No google image (unlike plot_map) so faster function with smaller output files.
}
