% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uff.r
\name{uff}
\alias{uff}
\title{Unique_flight_finder}
\usage{
uff(filenames, output = c("start.time", "flight.num"))
}
\arguments{
\item{filenames}{enter filenames which you want to find flight specific information about}

\item{output}{start times?}
}
\description{
Unique_flight_finder
}
